# Add performance tracing