--- a/src/launchpad/size/analyzers/apple.py
+++ b/src/launchpad/size/analyzers/apple.py
@@ -35,7 +35,7 @@ from launchpad.size.utils.apple_bundle_size import calculate_bundle_sizes
 from launchpad.utils.apple.code_signature_validator import CodeSignatureValidator
 from launchpad.utils.file_utils import calculate_file_hash, get_file_size
 from launchpad.utils.logging import get_logger
-from launchpad.utils.performance import trace_context_with_registry, trace_with_registry
+from launchpad.utils.performance import trace, trace_ctx
 
 from ..models.apple import (
     AppleAnalysisResults,
@@ -82,7 +82,7 @@ class AppleAppAnalyzer:
         self.skip_insights = skip_insights
         self.app_info: AppleAppInfo | None = None
 
-    @trace_with_registry("apple.preprocess")
+    @trace("apple.preprocess")
     def preprocess(self, artifact: AppleArtifact) -> AppleAppInfo:
         if not isinstance(artifact, ZippedXCArchive):
             raise NotImplementedError(f"Only ZippedXCArchive artifacts are supported, got {type(artifact)}")
@@ -90,7 +90,7 @@ class AppleAppAnalyzer:
         self.app_info = self._extract_app_info(artifact)
         return self.app_info
 
-    @trace_with_registry("apple.analyze")
+    @trace("apple.analyze")
     def analyze(self, artifact: AppleArtifact) -> AppleAnalysisResults:
         """Analyze an Apple app bundle.
 
@@ -188,7 +188,7 @@ class AppleAppAnalyzer:
 
         return results
 
-    @trace_with_registry("apple.extract_app_info")
+    @trace("apple.extract_app_info")
     def _extract_app_info(self, xcarchive: ZippedXCArchive) -> AppleAppInfo:
         """Extract basic app information.
 
@@ -234,7 +234,7 @@ class AppleAppAnalyzer:
             code_signature_errors=code_signature_errors,
         )
 
-    @trace_with_registry("apple.detect_file_type")
+    @trace("apple.detect_file_type")
     def _detect_file_type(self, file_path: Path) -> str:
         """Detect file type using the file command.
 
@@ -310,7 +310,7 @@ class AppleAppAnalyzer:
         # If no devices are provisioned, it's an app store profile
         return "appstore", profile_name
 
-    @trace_with_registry("apple.analyze_files")
+    @trace("apple.analyze_files")
     def _analyze_files(self, xcarchive: ZippedXCArchive) -> FileAnalysis:
         """Analyze all files in the app bundle.
 
@@ -380,7 +380,7 @@ class AppleAppAnalyzer:
 
         return FileAnalysis(files=files)
 
-    @trace_with_registry("apple.analyze_asset_catalog")
+    @trace("apple.analyze_asset_catalog")
     def _analyze_asset_catalog(self, xcarchive: ZippedXCArchive, relative_path: Path) -> List[TreemapElement]:
         """Analyze an asset catalog file."""
         catalog_details = xcarchive.get_asset_catalog_details(relative_path)
@@ -408,10 +408,10 @@ class AppleAppAnalyzer:
     def _generate_insight_with_tracing(
         self, insight_class: type, insights_input: InsightsInput, insight_name: str
     ) -> Any:
-        with trace_context_with_registry(f"apple.insights.{insight_name}"):
+        with trace_ctx(f"apple.insights.{insight_name}"):
             return insight_class().generate(insights_input)
 
-    @trace_with_registry("apple.analyze_binary")
+    @trace("apple.analyze_binary")
     def _analyze_binary(
         self, binary_path: Path, dwarf_binary_path: Path | None = None, skip_swift_metadata: bool = False
     ) -> MachOBinaryAnalysis:
@@ -499,7 +499,7 @@ class AppleAppAnalyzer:
             objc_method_names=objc_method_names,
         )
 
-    @trace_with_registry("apple.test_strip_symbols_removal")
+    @trace("apple.test_strip_symbols_removal")
     def _test_strip_symbols_removal(self, parser: MachOParser, binary_path: Path) -> int:
         """Test actual symbol removal using LIEF to get real size savings, similar to what strip does."""
         import tempfile