I'm noticing a couple bottlenecks when testing larger apps so wanted an easy way to see what's happening. Later on we can sync this with Sentry Tracing too and get production data.

```
           INFO     ============================================================                                     
           INFO     HIERARCHICAL PERFORMANCE SUMMARY                                                                 
           INFO     ============================================================                                     
           INFO     Total duration: 65.057s   Traces: 16   Executions: 95                                            
           INFO     ------------------------------------------------------------                                     
           INFO     └── apple.analyze: 65.057s (100.0 %)                                                             
           INFO         ├── apple.analyze_binary: 53.829s (82.7 %)  [16×, avg 3.364s]                                
           INFO         │   └── apple.test_strip_symbols_removal: 51.868s (79.7 %)  [16×, avg 3.242s]                
           INFO         ├── apple.analyze_files: 1.473s ( 2.3 %)                                                     
           INFO         │   ├── apple.detect_file_type: 0.755s ( 1.2 %)  [44×, avg 0.017s]                           
           INFO         │   └── apple.analyze_asset_catalog: 0.021s ( 0.0 %)  [7×, avg 0.003s]                       
           INFO         ├── apple.preprocess: 0.025s ( 0.0 %)                                                        
           INFO         │   └── apple.extract_app_info: 0.025s ( 0.0 %)                                              
           INFO         ├── apple.insights.duplicate_files: 0.000s ( 0.0 %)                                          
           INFO         ├── apple.insights.strip_binary: 0.000s ( 0.0 %)                                             
           INFO         ├── apple.insights.large_audio: 0.000s ( 0.0 %)                                              
           INFO         ├── apple.insights.large_images: 0.000s ( 0.0 %)                                             
           INFO         ├── apple.insights.large_videos: 0.000s ( 0.0 %)                                             
           INFO         ├── apple.insights.small_files: 0.000s ( 0.0 %)                                              
           INFO         ├── apple.insights.localized_strings: 0.000s ( 0.0 %)                                        
           INFO         └── apple.insights.hermes_debug_info: 0.000s ( 0.0 %)                                        
           INFO     ============================================================ 
```